+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## [1.0.0 Primera versión de la librería](https://svrgitpub.sdos.es/iOS/SDOSPluggableApplicationDelegate/tree/v1.0.0)
+## [2.0.0 Support Scenes](https://github.com/SDOSLabs/SDOSPluggableApplicationDelegate/tree/v2.0.0)
+
+- Support Scenes for iOS 13. New dependency `SDOSPluggableApplicationDelegate/Scene`
+
+## [1.0.0 Primera versión de la librería](https://github.com/SDOSLabs/SDOSPluggableApplicationDelegate/tree/v1.0.0)
 
 - Primera versión de la librería